scripts/requirements_hsbc_fund.txt
# HSBC Fund SQL Agent - 核心依赖
# 基于 LangChain SQL Agent 的自然语言转SQL系统

# Web 框架
fastapi==0.116.1
uvicorn>=0.24.0

# LangChain 核心组件 (兼容版本)
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.3
langchain-core==0.1.53

# 数据库连接 (PostgreSQL)
psycopg[binary]==3.2.9
SQLAlchemy>=2.0.0

# 版本约束 (解决兼容性问题)
pydantic>=2.0,<2.7

# 工具库
python-dotenv>=1.0.0
